move getjourney into seperate file repository
